module_utils/requirements.txt
collections
libvirt-python
